+++ Assets/Scripts/CorrectOrientation.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 866b091c94b63d5419818592b9f4f846
+guid: aced50e3fe99b8641b0fd149080eef1c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
